requirements.txt
flask
python-dotenv
flask-cors
unstructured
langchain
langchain-community
gunicorn
psutil
chromadb